User management is now up.
No support for password or login, and has been put aside, since relevance seems lacking at the current moment.

+++ VagtPlan.ApiService/DTOS/UserDTO.cs
@@ -4,8 +4,8 @@
     {
         public required string Name { get; set; }
         public required int VactionDays { get; set; }
-        public required int DepartmentId { get; set; }
-        public required int UserRoleId { get; set; }
+        public required long DepartmentId { get; set; }
+        public required long UserRoleId { get; set; }
     }
 
     public class LoginDTO
